1. add viewCount 2. add comment

+++ src/main/java/life/majiang/community/controller/PublishController.java
@@ -1,7 +1,6 @@
 package life.majiang.community.controller;
 
 import life.majiang.community.dto.*;
-import life.majiang.community.mapper.*;
 import life.majiang.community.model.*;
 import life.majiang.community.service.*;
 import org.springframework.beans.factory.annotation.*;
@@ -23,7 +22,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model) {
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title", question.getTitle());
@@ -43,7 +42,7 @@ public class PublishController {
             @RequestParam(name = "title", required = false) String title,
             @RequestParam(name = "description", required = false) String description,
             @RequestParam(name = "tag", required = false) String tag,
-            @RequestParam(name = "id", required = false) Integer id,
+            @RequestParam(name = "id", required = false) Long id,
             HttpServletRequest request,
             Model model) {
